third_party/WebKit/LayoutTests/http/tests/devtools/elements/edit/edit-style-attribute-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that style modification generates attribute updated event only when attribute is actually changed.


Running: testSetUp

Running: testSetNewValue
SDK.DOMModel.Events.AttrModified should be issued

Running: testSetSameValue
